requirements.txt
asgiref==3.2.2
Django==2.2.5
django-crispy-forms==1.8.0
django-tinymce4-lite==1.7.5
jsmin==2.2.2
Pillow==6.2.0
pytz==2019.2
sqlparse==0.3.0
